std: Remove `unsafe` from `ToCStr.{to,with}_c_str_unchecked`
It makes sense for ToCStr to have the default conversion to error out if the
string has null characters, but there are C apis out there that are written to
handle interior nulls in C strings.  These typically also take a length as part
of the api. We make sure that our CStrs always have a trailing null, so there's
no risk of a C function reading past the end of the string (unless they do it
on purpose). Given that I couldn't find a GCC or Clang warning that would
complain about "foo\0bar", I don't feel that the "_unchecked" variants should
be considered unsafe, just inappropriate in most cases.